add some benchmarks.
Performance isn't excellent in comparison to base - but this is expect
as I didn't implement it in a way that takes advantage of how
codepoints are assigned into places.

From now performance isn't a big concern for me but I'm curious why
GHC implements its lookup in C rather than Haskell and hopefully
having a benchmark offers me a bit of insight - it might also come in
handy as my future self might want to engineer some optimization.